requirements.txt
hydra-core==1.3.2
omegaconf==2.3.0
h5py==3.15.1
numcodecs==0.16.5
opencv-python==4.9.0.80
pandas==2.3.3
scipy==1.15.3
zarr==3.1.5
tqdm==4.67.1
einops==0.8.1
wandb==0.23.1
threadpoolctl==3.6.0
matplotlib==3.10.3
dill==0.4.0
pydantic==2.11.10
rtree==1.3.0
diffusers==0.35.2
pytorch-lightning==1.9.5
lightning-bolts==0.7.0
tyro==0.8.5
typing_extensions==4.12.2
numpy==1.26.0
sentry-sdk==2.29.1